move services to right packages
